Add files via upload
A universal function (or ufunc for short) is a function that operates on ndarrays in an element-by-element fashion
Example: np.add(2, x) where x is a Numpy array adds 2 to each element of the array.
Unary uFuncs vs. Binary uFuncs if they take one or two inputs
Example: np.add(2,x) vs. np.sum(x)
np.add() is used to add two arrays (or a scalar and an array) element-wise whereas np.sum() returns the sum of all elements in a single array.
ufuncs are used to implement vectorization in NumPy.
Vectorization refers to the use of pre-compiled functions and mathematical operations on Numpy array objects.
Vectorization also provide broadcasting and additional methods like reduce, accumulate etc. that are very helpful for computation
The alternative to vectorization would be to iterate through the elements of the array and perform the desired operations. Vectorization is much faster than iterating over elements. 
Unary uFuncs 
Simple element-wise transformations in a single array
Ex: np.sqrt(x) calculates the square root of each element in x
Arithmetic operations
Use symbol or equivalent uFunc
Binary uFuncs take two inputs
Aggregation Operations
Reduce or Accumulate methods
Statistical Aggregations (min, max, etc.)
Broadcasting operates on arrays of different shapes
The magic command (%timeit) gives the results in terms of averages and st.dev of multiple runs in nanoseconds (ns or 10^-9) or microseconds (us or 10^-6)
In contrast, the method timeit.timeit() gives a single estimate in seconds (as a float)
timeit.timeit(stmt, setup) executes the setup statement once, and returns the time it takes to execute the main statement (‘stmt’) a number of times, where number = 1000000
Since the number of runs is equal in all cases, the results are more comparable than if you use the magic command (%timeit) 
The disadvantage of the method is that it takes a bit more of code to produce. 
Check timeit advice for Python programmers at: 
https://www.infoworld.com/article/3532491/how-to-use-timeit-to-profile-python-code.html
Broadcasting means the ability to perform operations between arrays of different sizes.
Adding a number to a whole array is an example of broadcasting. 
Broadcasting is a feature that saves memory and coding, but it is somewhat difficult to understand at first.  
We will do a few exercises at the end of this session.
To learn more about Broadcasting and see examples consult the book chapter listed below.