Add files via upload
Pandas is a newer package built on top of NumPy
and provides an efficient implementation of a
DataFrame.
• DataFrames are essentially multidimensional arrays
with attached row and column labels, and often
with heterogeneous types and/or missing data.
• Import Pandas under the alias pd after importing NumPy
because of their dependency
Pandas objects can be thought of as enhanced
versions of NumPy structured arrays in which the
rows and columns are identified with labels rather
than simple integer indices.
• There are two main objects: Series and DataFrames
• Think of DataFrames as spreadsheets with more flexible
row and column labels
A Pandas Series is a one-dimensional array of indexed data. It can be created from a list or array
• The main difference between a NumPy 1D array and a
Panda series is the presence of the index
• while the NumPy array has an implicitly defined integer index used to access the values, the Pandas Series has an explicitly defined index associated with the values.
• Think of a Pandas Series a bit like a specialization of a
Python dictionary.
• A dictionary is a structure that maps arbitrary keys to a set of
arbitrary values, and a Series is a structure that maps typed
keys to a set of typed values.
• Different ways of creating a pd.series (Ex. 1)
A Pandas Series can be thought of as a single column in a
spreadsheet.
• For ease of representation, Pandas Series are sometimes
displayed horizontally because they have only one
dimension.
• The data type within a Pandas Series does not have to
homogeneous (i.e., same type).
• You can mix numeric and non-numeric data types within the same
series. Also, a series can have missing values (None)
Index object is an important structure in itself
• It can be thought of either as an immutable array or as an
ordered set
• Index as immutable array
• The Index object in many ways operates like an array. We can
use standard Python indexing notation to retrieve values or
slices
• One difference between Index objects and NumPy arrays is
that indices cannot be modified via the normal means (i.e.,
indexes are immutable)
• Index as ordered set
• The Index object follows many of the conventions used by
Python’s built-in set data structure, so that unions,
intersections, differences, and other combinations can be
computed in a familiar way
• Indexing Conventions
• Slice like in lists using the implicit index
• Fancy indexing refers to the explicit index or label
• Note difference between implicit and explicit index
when slicing with respect to the upper bound of the
range.
• Special Index attributes:
• loc – always references the explicit index
• iloc – always references the implicit index
• Rule: “explicit is better than implicit”
• Note: An index object can have its own name