Add files via upload
Pandas Concat & Merge
Finish the first part of the pandas module
• Function to generate synthetic DataFrames
• make_df in two versions: long and compact
• Concatenate with pd.concat
• Datasets with shared column names
• Datasets with different column names
• Concatenate with pd.merge
• Types of joins (one-to-one, one-to-many, many-to-many)
• Merge when the key column has different names
• Merge by keyword, or by index (Join)
• Read json data into a Pandas DataFrame
 Write a Python function to generate a
DataFrame, named make_df • Column labels are letters (A,B,C) and row indices are numbers
(0, 1,2). The location of the element (e.g., A1) is the value
• Pass as parameters string of letters (for column labels) and
range for index, return the newly created DataFrame
• Similar to concatenation of NumPy arrays via
np.concatenate
• Arguments is list of arrays to concatenate and axis
• Pandas function, pd.concat(), has a similar syntax
• By default, the concatenation takes place row-wise
within the DataFrame (i.e., axis=0)
• Difference between np.concatenate and pd.concat
is that Pandas concatenation preserves indices,
even if the result will have duplicate indices
pd.concat offers several options to join datasets
with different column names
• By default, the entries for which no data is available are
filled with NA values.
• To change default option, specify one of several options
for the join parameters
• the join default is a union of the input columns (join=
'outer'), but it can be changed to the intersection of the
columns using join='inner’
The pd.merge() looks for one or more
matching column names between the two
inputs, and uses this as the key
• It implements different types of joins
depending on the form of the input data.
• One-to-one join is in similar to column-wise
concatenation
• Many-to-one joins are joins in which one of the
two key columns contains duplicate entries.
The resulting DataFrame will preserve those
duplicate entries as appropriate.
• Many-to-many joins happen if the key column
in both the left and right array contains
duplicates.
• To specify the key, use the on keyword
• This only works if the left and right DataFrame have the specified column name
• To merge two datasets with different column names, use
the left_on and right_on keywords to specify the two column names
• The result has a redundant column that we can drop if desired
using the drop() method
• To merge on an index, instead of merging on a column, use the right_index and left_index. • For convenience, DataFrames implement the join() method, which performs a merge that defaults to joining on indices
• To mix indices and columns, you can combine on and index
• For example, left_index with right_on or left_on with right_index to
get the desired behavior
What happens when a value appears in one key
column but not the other?
• By default, the result contains the intersection of
the two sets of inputs, known as an inner join
• This can be specified explicitly using the how keyword
(which defaults to 'inner’)
• An outer join returns a join over the union of the
input columns, and fills missing values with NAs
• The left join and right join return join over the left
entries and right entries, respectively
Json stands for Java Script Object Notation
• Name was inspired by the syntax of JavaScript (a
programming language used for web development).
However, it has become a language-independent data
format.
• Json format is widely used in web applications and
configuration files
• Main characteristics of the JSON format:
• There is a sequence of key-value pairs surrounded by curly
brackets {}.
• Each key is mapped to a particular value using this format:
• Key-value pairs are separated by a comma. Only the last pair
is not followed by a comma.
• JSON and Dictionaries: visually similar but different
• JSON is a file format used to represent and store data
whereas a Python Dictionary is the actual data structure
(object) that is kept in memory while a Python program
runs.
• The key-value pairs of the JSON file can be used to create a
Python dictionary with the help of special libraries.
• JSON is the string representation of the data and
dictionaries are the actual data structures in memory that
are created when the program runs.
pd.read_json (data, orient,…)
• is used to Convert a VALID JSON
string to a Pandas object.
• The parameter orient is used to
align index and columns.
• For DataFrame, the default is columns
whereas other allowed values are
split, record, index, values and table.
• The string could be a file handle or
a URL to a json file.