Add files via upload
DATA AGGREGATION & ANALYSIS!!!
• Once the data is in a pandas DataFrame, the next
step is analyzing it to answer some questions or
extract some insights.
• GroupBy allows simple aggregation by groups
based on values or based on a condition.
• Depending on the case, GroupBy involves one or
more steps of the Split-Combine-Apply paradigm.
The split step involves breaking up and grouping a
DataFrame depending on the value of the specified
key.
• The apply step involves computing some function,
usually an aggregate,
transformation, or filtering,
within the individual
groups.
• The combine step merges the results of these
operations into an output
array.
When you use the GroupBy operation in a
DataFrame, Python returns a GroupBy object
which is an abstraction, or a special view of the
DataFrame.
• This view is not a set of DataFrames, but it works as if it
was a collection of them. To calculate values, an
aggregation operation is needed.
• The DataFrame Group-By object is a flexible
abstraction that supports any valid DataFrame
operation.
• Iteration over groups
• Dispatch methods
These GroupBy object methods efficiently implement a
variety of useful operations before combining the grouped
data.
• Aggregation. The aggregate() method allows for more flexibility
than GroupBy aggregations with sum(), median(), and so on. It can
take a string, a function, or a list thereof, and compute all the
aggregates at once.
• Filtering. A filtering operation allows you to drop data based on the
group properties. For example, keep all groups in which the
standard deviation is larger than some critical value.
• Transformation. It is used to return some transformed version of
the full data to recombine. For such a transformation, the output is
the same shape as the input. An example of this would be to center
the data by subtracting the group-wise mean
• Apply. The apply() method lets you apply an arbitrary function to
the group results. The function takes a DataFrame and returns
either a Pandas object or a scalar
Uses of Lambda Functions
• To include a function into another that is named and declared with def
• To perform calculations embedded in code (e.g., create a new df column)
• To leverage calculations for filtering operations in a more efficient way