Add files via upload
DATA AGGREGATION & ANALYSIS : PIVOT TABLES 
We explored the planet dataset in the seaborn
library using GroupBy
• Challenge Question: How many planets were discovered
by method in each decade?
• For the pending challenge question, we need to
create new series (decade) for each row by
applying integer division (years // 10) and
multiplying by 10.
• Create a new column called ‘Decade’ in the dataset
The GroupBy abstraction allows us to explore relationships within a dataset.
• Limitations: uni-dimensionality and complexity
• A pivot table is a similar operation with streamlined parameters
that allows more than one aggregation dimension.
• Pivot Tables are commonly seen in spreadsheets and other programs
that operate on tabular data.
• The phrase Pivot Table was trademarked by Microsoft in the 1990s.
• A pivot table takes simple column-wise data as input, and groups the entries into a two-dimensional table that provides a multidimensional summarization of the data.
• The difference between pivot tables and GroupBy can sometimes
cause confusion. Think of pivot tables as essentially a
multidimensional version of GroupBy aggregation.
• That is, you split-apply-combine, but both the split and the combine happen across not a one-dimensional index, but across a two- dimensional grid.
Data is the DataFrame and values is the column to
aggregate (optional)
• The index and column keywords are column, grouper, array or list to aggregate.
• The aggfunc keyword controls what type of aggregation is applied. The default is the mean.
• As in the GroupBy, the aggregation specification can be a string representing one of several common choices ('sum',
'mean', 'count', 'min', 'max', etc.)
• The fill_value and dropna have to do with handling missing data
• The margin keyword allows to compute totals along each grouping. 
For the pivot tables code exercises, we will use the
Titanic dataset available in the seaborn package.
• The first step after you import the dataset is to
familiarize yourself with the shape, columns and
row, and data types.
The Pandas CrossTab method allows you to build
a cross-tabulation table to show the frequency
with which certain groups of data appear.
• Imagine a cross-tabulation as a bidimensional
grid where the cells show the counts of elements
with the characteristics indicated in the
intersection of the row and the column
• In other words, the joint frequency of the two
categorical variables that define the row and column
groupings.
GroupBy is useful to create abstractions (views)
of the data in one-dimensional categories.
• Pivot Tables are used to run analyses on more
than one dimension, and they are the main tool
for most of the data analysis needs.
• CrossTab method allows you to build a crosstabulation table to show the frequency with
which certain groups of data appear.